feat[st]: Factory for data of challenge projects
