+++ SmartTutor/Repository/ActiveEducationRepository/ActivityFactory.cs
@@ -1,7 +1,7 @@
 ﻿using LibGit2Sharp;
-using RepositoryCompiler.CodeModel;
 using RepositoryCompiler.RepositoryAdapters;
 using SmartTutor.ActiveEducationModel;
+using SmartTutor.Repository.ChallengeProjectRepository;
 using System;
 using System.Collections.Generic;
 
@@ -9,6 +9,8 @@ namespace SmartTutor.Repository.ActiveEducationRepository
 {
     public class ActivityFactory
     {
+        private readonly ChallengeProjectFactory challengeProjectFactory;
+
         public Dictionary<SmellType, List<EducationActivity>> CreateActivities()
         {
             Dictionary<SmellType, List<EducationActivity>> educationActivities = new Dictionary<SmellType, List<EducationActivity>>();
@@ -55,26 +57,10 @@ namespace SmartTutor.Repository.ActiveEducationRepository
                     Rank = 1,
                     EducationActivities = new List<EducationActivity>()
                 },
-                Project = CreateLongMethodProject()
+                Project = challengeProjectFactory.CreateLongMethodProject()
             };
 
             return longMethodChallenge;
-        }
-
-        public ChallengeProject CreateLongMethodProject()
-        {
-            return new ChallengeProject
-            {
-                Name = "Extract AwardAchievement method",
-                Description = "1) Discover (e.g., using Google) the Extract Method command in your IDE." +
-                                " 2) Using the command, extract multiple methods from the AwardAchievement method." +
-                                " 3) For each method, define the most appropriate name.",
-                Level = 1,
-                Points = 5,
-                GitURL = "https://github.com/Ana00000/Challenge-inspiration.git",
-                StartState = new CodeModelFactory(LanguageEnum.CSharp).CreateProjectWithCodeFileLinks(@"C:\ActiveEducation\SRP\SRP1"),
-                EndState = new CodeModelFactory(LanguageEnum.CSharp).CreateProjectWithCodeFileLinks(@"C:\ActiveEducation\SRP\SRP1")
-            };
         }
 
         private List<EducationTraining> CreateLongMethodTrainings()
